toolkit/components/extensions/test/mochitest/test_ext_webrequest.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for simple WebExtension</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

SimpleTest.requestCompleteLog();

const BASE = "http://mochi.test:8888/tests/toolkit/components/extensions/test/mochitest";

const expected_requested = [BASE + "/file_WebRequest_page1.html",
                            BASE + "/file_style_good.css",
                            BASE + "/file_style_bad.css",
                            BASE + "/file_style_redirect.css",
                            BASE + "/file_image_good.png",
                            BASE + "/file_image_bad.png",
                            BASE + "/file_image_redirect.png",
                            BASE + "/file_script_good.js",
                            BASE + "/file_script_bad.js",
                            BASE + "/file_script_redirect.js",
                            BASE + "/file_script_xhr.js",
                            BASE + "/file_WebRequest_page2.html",
                            BASE + "/nonexistent_script_url.js",
                            BASE + "/redirection.sjs",
                            BASE + "/xhr_resource"];

const expected_beforeSendHeaders = [BASE + "/file_WebRequest_page1.html",
                              BASE + "/file_style_good.css",
                              BASE + "/file_style_redirect.css",
                              BASE + "/file_image_good.png",
                              BASE + "/file_image_redirect.png",
                              BASE + "/file_script_good.js",
                              BASE + "/file_script_redirect.js",
                              BASE + "/file_script_xhr.js",
                              BASE + "/file_WebRequest_page2.html",
                              BASE + "/nonexistent_script_url.js",
                              BASE + "/redirection.sjs",
                              BASE + "/dummy_page.html",
                              BASE + "/xhr_resource"];

const expected_sendHeaders = expected_beforeSendHeaders.filter(u => !/_redirect\./.test(u))
                            .concat(BASE + "/redirection.sjs");

const expected_redirect = expected_beforeSendHeaders.filter(u => /_redirect\./.test(u))
                            .concat(BASE + "/redirection.sjs");

const expected_complete = [BASE + "/file_WebRequest_page1.html",
                           BASE + "/file_style_good.css",
                           BASE + "/file_image_good.png",
                           BASE + "/file_script_good.js",
                           BASE + "/file_script_xhr.js",
                           BASE + "/file_WebRequest_page2.html",
                           BASE + "/nonexistent_script_url.js",
                           BASE + "/dummy_page.html",
                           BASE + "/xhr_resource"];

function removeDupes(list) {
  let j = 0;
  for (let i = 1; i < list.length; i++) {
    if (list[i] != list[j]) {
      j++;
      if (i != j) {
        list[j] = list[i];
      }
    }
  }
  list.length = j + 1;
}

function compareLists(list1, list2, kind) {
  list1.sort();
  removeDupes(list1);
  list2.sort();
  removeDupes(list2);
  is(String(list1), String(list2), `${kind} URLs correct`);
}

function backgroundScript() {
  const BASE = "http://mochi.test:8888/tests/toolkit/components/extensions/test/mochitest";

  let checkCompleted = true;
  let savedTabId = -1;

  function checkType(details) {
    let expected_type = "???";
    if (details.url.indexOf("style") != -1) {
      expected_type = "stylesheet";
    } else if (details.url.indexOf("image") != -1) {
      expected_type = "image";
    } else if (details.url.indexOf("script") != -1) {
      expected_type = "script";
    } else if (details.url.indexOf("page1") != -1) {
      expected_type = "main_frame";
    } else if (/page2|redirection|dummy_page/.test(details.url)) {
      expected_type = "sub_frame";
    } else if (details.url.indexOf("xhr") != -1) {
      expected_type = "xmlhttprequest";
    }
    browser.test.assertEq(details.type, expected_type, "resource type is correct");
  }

  let frameIDs = new Map();

  let recorded = {requested: [],
                  beforeSendHeaders: [],
                  beforeRedirect: [],
                  sendHeaders: [],
                  responseStarted: [],
                  completed: []};

  function checkResourceType(type) {
    let key = type.toUpperCase();
    browser.test.assertTrue(key in browser.webRequest.ResourceType);
  }

  function onBeforeRequest(details) {
    browser.test.log(`onBeforeRequest ${details.url}`);
    checkResourceType(details.type);
    if (details.url.startsWith(BASE)) {
      recorded.requested.push(details.url);

      if (savedTabId == -1) {
        browser.test.assertTrue(details.tabId !== undefined, "tab ID defined");
        savedTabId = details.tabId;
      }

      browser.test.assertEq(details.tabId, savedTabId, "correct tab ID");
      checkType(details);

      frameIDs.set(details.url, details.frameId);
      if (details.url.indexOf("page1") != -1) {
        browser.test.assertEq(details.frameId, 0, "frame ID correct");
        browser.test.assertEq(details.parentFrameId, -1, "parent frame ID correct");
      }
      if (details.url.indexOf("page2") != -1) {
        browser.test.assertTrue(details.frameId != 0, "sub-frame gets its own frame ID");
        browser.test.assertTrue(details.frameId !== undefined, "sub-frame ID defined");
        browser.test.assertEq(details.parentFrameId, 0, "parent frame id is correct");
      }
    }
    if (details.url.indexOf("_bad.") != -1) {
      return {cancel: true};
    }
    return {};
  }

  function onBeforeSendHeaders(details) {
    browser.test.log(`onBeforeSendHeaders ${details.url}`);
    checkResourceType(details.type);
    if (details.url.startsWith(BASE)) {
      recorded.beforeSendHeaders.push(details.url);

      browser.test.assertEq(details.tabId, savedTabId, "correct tab ID");
      checkType(details);

      let id = frameIDs.get(details.url);
      browser.test.assertEq(id, details.frameId, "frame ID same in onBeforeSendHeaders as onBeforeRequest");
    }
    if (details.url.indexOf("_redirect.") != -1) {
      return {redirectUrl: details.url.replace("_redirect.", "_good.")};
    }
    return {};
  }

  function onBeforeRedirect(details) {
    browser.test.log(`onBeforeRedirect ${details.url} -> ${details.redirectUrl}`);
    checkResourceType(details.type);
    if (details.url.startsWith(BASE)) {
      recorded.beforeRedirect.push(details.url);

      browser.test.assertEq(details.tabId, savedTabId, "correct tab ID");
      checkType(details);

      let id = frameIDs.get(details.url);
      browser.test.assertEq(id, details.frameId, "frame ID same in onBeforeRedirect as onBeforeRequest");
      frameIDs.set(details.redirectUrl, details.frameId);
    }
    if (details.url.indexOf("_redirect.") != -1) {
      let expectedUrl = details.url.replace("_redirect.", "_good.");
      browser.test.assertEq(details.redirectUrl, expectedUrl, "correct redirectUrl value");
    }
    return {};
  }

  function onRecord(kind, details) {
    checkResourceType(details.type);
    if (details.url.startsWith(BASE)) {
      recorded[kind].push(details.url);
    }
  }

  let completedUrls = {
    responseStarted: new Set(),
    completed: new Set(),
  };

  function checkIpAndRecord(kind, details) {
    onRecord(kind, details);

    // When resources are cached, the ip property is not present,
    // so only check for the ip property the first time around.
    if (checkCompleted && !completedUrls[kind].has(details.url)) {
      browser.test.assertEq(details.ip, "127.0.0.1", "correct ip");
      completedUrls[kind].add(details.url);
    }
  }

  browser.webRequest.onBeforeRequest.addListener(onBeforeRequest, {urls: ["<all_urls>"]}, ["blocking"]);
  browser.webRequest.onBeforeSendHeaders.addListener(onBeforeSendHeaders, {urls: ["<all_urls>"]}, ["blocking"]);
  browser.webRequest.onSendHeaders.addListener(onRecord.bind(null, "sendHeaders"), {urls: ["<all_urls>"]});
  browser.webRequest.onBeforeRedirect.addListener(onBeforeRedirect, {urls: ["<all_urls>"]});
  browser.webRequest.onResponseStarted.addListener(checkIpAndRecord.bind(null, "responseStarted"), {urls: ["<all_urls>"]});
  browser.webRequest.onCompleted.addListener(checkIpAndRecord.bind(null, "completed"), {urls: ["<all_urls>"]});

  function onTestMessage(msg) {
    if (msg == "skipCompleted") {
      checkCompleted = false;
      browser.test.sendMessage("ackSkipCompleted");
    } else {
      browser.test.sendMessage("results", recorded);
    }
  }

  browser.test.onMessage.addListener(onTestMessage);

  browser.test.sendMessage("ready", browser.webRequest.ResourceType);
}

function* test_once(skipCompleted) {
  let extensionData = {
    manifest: {
      permissions: [
        "webRequest",
        "webRequestBlocking",
      ],
    },
    background: "(" + backgroundScript.toString() + ")()",
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);
  let [, resourceTypes] = yield Promise.all([extension.startup(), extension.awaitMessage("ready")]);
  info("webrequest extension loaded");

  if (skipCompleted) {
    extension.sendMessage("skipCompleted");
    yield extension.awaitMessage("ackSkipCompleted");
  }

  for (let key in resourceTypes) {
    let value = resourceTypes[key];
    is(key, value.toUpperCase());
  }

  // Check a few Firefox-specific types.
  is(resourceTypes.XBL, "xbl", "XBL resource type supported");
  is(resourceTypes.FONT, "font", "Font resource type supported");
  is(resourceTypes.WEBSOCKET, "websocket", "Websocket resource type supported");

  yield new Promise(resolve => { setTimeout(resolve, 0); });

  let win = window.open();

  // Clear the image cache, since it gets in the way otherwise.
  let imgTools = SpecialPowers.Cc["@mozilla.org/image/tools;1"].getService(SpecialPowers.Ci.imgITools);
  let cache = imgTools.getImgCacheForDocument(win.document);
  cache.clearCache(false);

  // yield waitForLoad(win);
  info("about:blank loaded");

  win.location = "file_WebRequest_page1.html";

  yield waitForLoad(win);
  info("test page loaded");

  is(win.success, 2, "Good script ran");
  is(win.failure, undefined, "Failure script didn't run");

  let style = win.getComputedStyle(win.document.getElementById("test"), null);
  is(style.getPropertyValue("color"), "rgb(255, 0, 0)", "Good CSS loaded");

  win.close();

  extension.sendMessage("getResults");
  let recorded = yield extension.awaitMessage("results");

  compareLists(recorded.requested, expected_requested, "requested");
  compareLists(recorded.beforeSendHeaders, expected_beforeSendHeaders, "beforeSendHeaders");
  compareLists(recorded.sendHeaders, expected_sendHeaders, "sendHeaders");
  compareLists(recorded.beforeRedirect, expected_redirect, "beforeRedirect");
  compareLists(recorded.responseStarted, expected_complete, "responseStarted");
  compareLists(recorded.completed, expected_complete, "completed");

  yield extension.unload();
  info("webrequest extension unloaded");
}

// Run the test twice to make sure it works with caching.
add_task(function*() { yield test_once(false); });
add_task(function*() { yield test_once(true); });
</script>

</body>
</html>
